refactor(serve): split placement.py from subprocess_mgr, move tunables to constants, spawn as classmethod
Three related cleanups on top of the prior PR 3 commit:

1. subprocess_mgr.py lost its placement-group half to a new
   placement.py. subprocess_mgr is now strictly ManagedSubprocess +
   the _SubprocessActor factory + graceful_stop_actors primitive;
   placement.py owns ReplicaBundleSpec, the planner
   (plan_replica_bundle_shape, _get_gpu_topology),
   build_pg / build_replica_pg, the bundle-scoped discovery helpers
   (get_bundle_node_ip, get_free_port_in_bundle, _run_in_bundle),
   and remove_named_pgs_with_prefix. Concrete win: dynamo/infra.py
   imports build_pg from placement now, which reads correctly;
   previously it pulled it out of subprocess_mgr which was wrong
   name-wise. The Bundle(pg, idx) wrapper TODO lives with the bundle
   helpers in placement.py.

2. The module-level tunables that used to sit at the top of
   subprocess_mgr moved to nemo_curator/core/serve/constants.py and
   dropped the underscore prefix -- they were being imported across
   files (dynamo/infra.py pulled PG_READY_TIMEOUT_S and
   WORKER_NODE_LABEL), so "private" was misleading. Same file now
   holds SIGTERM_WAIT_S, SIGKILL_WAIT_S,
   PLACEMENT_GROUP_READY_TIMEOUT_S (renamed from PG_... for
   readability), WORKER_NODE_LABEL, and NOSET_CUDA_RUNTIME_ENV,
   alongside the existing DEFAULT_SERVE_PORT /
   DEFAULT_SERVE_HEALTH_TIMEOUT_S.

3. spawn_actor free function is gone; spawning is now
   ManagedSubprocess.spawn(label, pg, bundle_index, ...) classmethod,
   so the factory lives on the type it returns. The previously
   underscore-private _build_pg is now just build_pg (public; it's
   imported across modules).

Tests tracked the split:
  - tests/core/serve/test_placement.py (new) -- planner matrix,
    head-node exclusion, ReplicaBundleSpec, orphan PG cleanup.
  - tests/core/serve/test_subprocess_mgr.py narrowed to the
    lifecycle coverage: TestReplicaLifecycle (end-to-end spawn +
    env propagation + wait + read_log_tail + stop) and
    test_actor_death_surfaces_via_run_ref.

Verified on 2 GPUs:
  - 21/21 non-GPU tests pass (placement planner + dynamo/infra + test_infra + test_config + test_server, etc.)
  - 3/3 GPU tests pass (TestReplicaLifecycle.test_end_to_end,
    test_actor_death_surfaces_via_run_ref, test_orphan_pg_cleanup_by_prefix)
